Adapter: Move dataBlob persistence to datasource package
