+++ adapter/src/main/java/org/jvalue/ods/adapterservice/adapter/api/rest/v1/DataEndpoint.java
@@ -1,6 +1,6 @@
 package org.jvalue.ods.adapterservice.adapter.api.rest.v1;
 
-import org.jvalue.ods.adapterservice.adapter.DataBlobRepository;
+import org.jvalue.ods.adapterservice.datasource.repository.DataBlobRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.GetMapping;
