+++ adapter/src/main/java/org/jvalue/ods/adapterservice/datasource/repository/DataBlobRepository.java
@@ -1,6 +1,6 @@
-package org.jvalue.ods.adapterservice.adapter;
+package org.jvalue.ods.adapterservice.datasource.repository;
 
-import org.jvalue.ods.adapterservice.adapter.model.DataBlob;
+import org.jvalue.ods.adapterservice.datasource.model.DataBlob;
 import org.springframework.data.repository.CrudRepository;
 
 public interface DataBlobRepository extends CrudRepository<DataBlob, Long> {
